Silence a lot of warnings

+++ id_pm.c
@@ -78,7 +78,7 @@ void PM_Startup (void)
         if (!pageOffsets[i])
             continue;           // sparse page
 
-        if (pageOffsets[i] < pageOffsets[0] || pageOffsets[i] >= filesize)
+        if (pageOffsets[i] < pageOffsets[0] || pageOffsets[i] >= (size_t)filesize)
             Quit ("PM_Startup: Illegal page offset for page %i: %u (filesize: %u)",i,pageOffsets[i],filesize);
     }
